+++ instrumentation/github.com/gorilla/mux/otelmux/mux.go
@@ -12,13 +12,13 @@ import (
 	"github.com/gorilla/mux"
 
 	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/request"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/semconv"
 	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/semconvutil"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/propagation"
-	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -55,7 +55,6 @@ func Middleware(service string, opts ...Option) mux.MiddlewareFunc {
 		ScopeName,
 		metric.WithInstrumentationVersion(Version()),
 	)
-
 	return func(handler http.Handler) http.Handler {
 		return traceware{
 			service:           service,
@@ -66,7 +65,7 @@ func Middleware(service string, opts ...Option) mux.MiddlewareFunc {
 			publicEndpoint:    cfg.PublicEndpoint,
 			publicEndpointFn:  cfg.PublicEndpointFn,
 			filters:           cfg.Filters,
-			meter:             meter,
+			semconv:           semconv.NewHTTPServer(meter),
 		}
 	}
 }
@@ -81,6 +80,7 @@ type traceware struct {
 	publicEndpointFn  func(*http.Request) bool
 	filters           []Filter
 	meter             metric.Meter
+	semconv           semconv.HTTPServer
 
 	requestBytesCounter  metric.Int64Counter
 	responseBytesCounter metric.Int64Counter
@@ -111,21 +111,8 @@ func (tw traceware) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	tw.createMeasures()
 	ctx := tw.propagators.Extract(r.Context(), propagation.HeaderCarrier(r.Header))
-	routeStr := ""
-	route := mux.CurrentRoute(r)
-	if route != nil {
-		var err error
-		routeStr, err = route.GetPathTemplate()
-		if err != nil {
-			routeStr, err = route.GetPathRegexp()
-			if err != nil {
-				routeStr = ""
-			}
-		}
-	}
-
 	opts := []trace.SpanStartOption{
-		trace.WithAttributes(semconvutil.HTTPServerRequest(tw.service, r)...),
+		trace.WithAttributes(tw.semconv.RequestTraceAttrs(tw.service, r)...),
 		trace.WithSpanKind(trace.SpanKindServer),
 	}
 
@@ -137,14 +124,22 @@ func (tw traceware) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
+	routeStr := ""
+	route := mux.CurrentRoute(r)
+	if route != nil {
+		routeStr, _ = route.GetPathTemplate()
+		if routeStr == "" {
+			routeStr, _ = route.GetPathRegexp()
+		}
+	}
+
 	if routeStr == "" {
 		routeStr = fmt.Sprintf("HTTP %s route not found", r.Method)
 	} else {
-		rAttr := semconv.HTTPRoute(routeStr)
+		rAttr := tw.semconv.Route(routeStr)
 		opts = append(opts, trace.WithAttributes(rAttr))
 	}
-	spanName := tw.spanNameFormatter(routeStr, r)
-	ctx, span := tw.tracer.Start(ctx, spanName, opts...)
+	ctx, span := tw.tracer.Start(ctx, tw.spanNameFormatter(routeStr, r), opts...)
 	defer span.End()
 
 	readRecordFunc := func(int64) {}
@@ -179,10 +174,14 @@ func (tw traceware) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	tw.handler.ServeHTTP(w, r.WithContext(ctx))
 	statusCode := rww.StatusCode()
-	if statusCode > 0 {
-		span.SetAttributes(semconv.HTTPStatusCode(statusCode))
-	}
-	span.SetStatus(semconvutil.HTTPServerStatus(statusCode))
+	span.SetStatus(tw.semconv.Status(statusCode))
+	span.SetAttributes(tw.semconv.ResponseTraceAttrs(semconv.ResponseTelemetry{
+		StatusCode: statusCode,
+		ReadBytes:  bw.BytesRead(),
+		ReadError:  bw.Error(),
+		WriteBytes: rww.BytesWritten(),
+		WriteError: rww.Error(),
+	})...)
 
 	// Add metrics
 	attributes := append([]attribute.KeyValue{}, semconvutil.HTTPServerRequestMetrics(tw.service, r)...)